+++ tango/factory.go
@@ -3,10 +3,15 @@ package tango
 import (
 	"github.com/tyler-sommer/stick"
 	"github.com/tyler-sommer/stick/twig/filter"
+	"os"
 )
 
+// CreateStick creates a new stick.Env with the default filters and the filters defined in this package
+// It also adds the current working directory to the filesystem loader
 func CreateStick() *stick.Env {
-	env := stick.New(nil)
+	d, _ := os.Getwd()
+	loader := stick.NewFilesystemLoader(d)
+	env := stick.New(loader)
 	env.Filters = mergeMaps(filter.TwigFilters(), CreateFilters())
 	return env
 }